rebuild: clean up rebuild job on child removal
when a child is being remove we check for any existing jobs that has:
1. child as dest
in this case we simply stop the job and wait for it to finish
2. child as src
same as above but we also need to start a new rebuild job to rebuild the
dest child.

Atm the wait infra is not here so we have to rely on the child destroy
to block until all the bdevs are closed.
This will be addressed once cas-194 is resolved.

CAS-126